requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
click==8.1.3
python-dateutil==2.8.2
six==1.16.0
flask-wtf==1.1.1
wtforms==3.0.1
python-dotenv==1.0.0
reportlab==3.6.12
pillow==9.5.0
colorama==0.4.6
